+++ defmt/src/impls/alloc_.rs
@@ -44,3 +44,23 @@ where
 impl<'a> Format for alloc::borrow::Cow<'a, str> {
     delegate_format!(str, self, self.as_ref());
 }
+
+impl Format for alloc::string::FromUtf8Error {
+    delegate_format!(core::str::Utf8Error, self, &self.utf8_error());
+}
+
+impl Format for alloc::ffi::IntoStringError {
+    fn format(&self, fmt: Formatter) {
+        crate::write!(fmt, "C string contained non-utf8 bytes")
+    }
+}
+
+impl Format for alloc::ffi::NulError {
+    fn format(&self, fmt: Formatter) {
+        crate::write!(
+            fmt,
+            "nul byte found in provided data at position: {}",
+            self.nul_position()
+        )
+    }
+}