+++ CHANGELOG.md
@@ -62,6 +62,7 @@ We have several packages which live in this repository. Changes are tracked sepa
 
 * [#1089] Retain timestamp and bitflags metadata when linking without `defmt.x`.
 * [#1068] Adding `Format` impl for `core::str` errors
+* [#1096] Implement `Format` for `alloc` Errors
 
 ### [defmt-v1.1.1] (2026-06-26)
 
@@ -1042,6 +1043,7 @@ Initial release
 
 ---
 
+[#1096]: https://github.com/knurling-rs/defmt/pull/1096
 [#1089]: https://github.com/knurling-rs/defmt/pull/1089
 [#1084]: https://github.com/knurling-rs/defmt/pull/1084
 [#1083]: https://github.com/knurling-rs/defmt/pull/1083